cognito_abac_authorizer/requirements.txt
cffi==1.16.0
cryptography==42.0.7
pycparser==2.22
PyJWT==2.8.0
